pyproject.toml
[project]
name = "route-master-africac-group7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-bcrypt>=1.0.1",
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.2.1",
]
